Module cfg_primitives::constants

Common constants for all runtimes

Modules

currency_decimals

Constants

AVERAGE_ON_INITIALIZE_RATIO

We assume that ~5% of the block weight is consumed by on_initialize handlers. This is used to limit the maximal weight of a single extrinsic.

CENTI_CFG

CFG

DAYS

DEFAULT_FEE_VALUE

Value for a not specified fee key.

HOURS

MAXIMUM_BLOCK_WEIGHT

We allow for 0.5 seconds of compute with a 6 second average block time.

MICRO_CFG

MILLISECS_PER_BLOCK

This determines the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

MILLISECS_PER_DAY

Milliseconds per day

MILLI_CFG

MINUTES

MIN_VESTING

Minimum vesting amount, in CFG/AIR

NFTS_PREFIX

These are pre/appended to the registry id before being set as a [RegistryInfo] field in [create_registry].

NORMAL_DISPATCH_RATIO

We allow Normal extrinsics to fill up the block up to 75%, the rest can be used by Operational extrinsics.

SECONDS_PER_DAY

SECONDS_PER_HOUR

SECONDS_PER_MINUTE

SECONDS_PER_YEAR

SLOT_DURATION

TREASURY_FEE_RATIO

% of fee addressed to the Treasury. The reminder % will be for the block author.

Functions

deposit