Module cfg_primitives::constants

Common constants for all runtimes

Modules

currency_decimals

Constants

AVERAGE_ON_INITIALIZE_RATIO

We assume that ~5% of the block weight is consumed by on_initialize handlers. This is used to limit the maximal weight of a single extrinsic.

CENTI_CFG

CFG

DAYS

DEFAULT_FEE_VALUE

Value for a not specified fee key.

GENERAL_CURRENCY_INDEX_PREFIX

Unhashed 36-bytes prefix for currencies managed by LiquidityPools.

HOURS

MAXIMUM_BLOCK_WEIGHT

We allow for 0.5 seconds of compute with a 6 second average block time.

MAX_TOKEN_NAME_LENGTH_BYTES

The max length allowed for a tranche token name

MAX_TOKEN_SYMBOL_LENGTH_BYTES

The max length allowed for a tranche token symbol

MICRO_CFG

MILLISECS_PER_BLOCK

This determines the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

MILLISECS_PER_DAY

Milliseconds per day

MILLI_CFG

MINUTES

MIN_VESTING

Minimum vesting amount, in CFG/AIR

NORMAL_DISPATCH_RATIO

We allow Normal extrinsics to fill up the block up to 75%, the rest can be used by Operational extrinsics.

SAFE_XCM_VERSION

The safe XCM version of pallet-xcm, same as on relay chain

SECONDS_PER_DAY

SECONDS_PER_HOUR

SECONDS_PER_MINUTE

SECONDS_PER_MONTH

SECONDS_PER_WEEK

SECONDS_PER_YEAR

SLOT_DURATION

TRANSACTION_RECOVERY_ID

Transaction recovery ID used for generating a signature in the Ethereum Transaction pallet. As per: https://github.com/PureStake/moonbeam/blob/fb63014a5e487f17e31283776e4f6b0befd009a2/primitives/xcm/src/ethereum_xcm.rs#L167

TREASURY_FEE_RATIO

% of fee addressed to the Treasury. The reminder % will be for the block author.

Functions

deposit